Use the 'base event' instead of 'first instance' when editing recurrent events
First instance on a recurrent event is not special. AgenDAV used it as a
template when editing all instances of a recurrent event, which is wrong, and
also introduced a rendering bug (#170).

Now the base event is fetched from the server (AgenDAV can't store events on the
frontend) without rendering it. This fixes #170.