Fix ansible SSL issue with github.com
See for example https://github.com/ansible/ansible/issues/12161

There are numerous ansible issues reported with the following message :
  msg: Failed to validate the SSL certificate for domain.tld:443. Use validate_certs=False or make sure your managed systems have a valid CA certificate installed. Paths checked for this platform: /etc/ssl/certs, /etc/pki/ca-trust/extracted/pem, /etc/pki/tls/certs, /usr/share/ca-certificates/cacert.org, /etc/ansible

It seems to be caused by python <2.7.9 not supporting SNI
Running the baikal playbook on ubuntu trusty, with default python-2.7 which is 2.7.6,
we run into this error.
As noted in the issue, using curl/wget works.