[BUGFIX] Update display dates timezone table
The library used to calculate dates and timezones is outdated.
As a result the latest switch to summertime in the EU results in
timezone offsets for all new events.

Update the timezone libary with latest timezone tables
to prevent offset after saving an event.

Note: This fix requires running `npm run build:js` afterwards
in order to rebuild assets in the `dist` directory.

Refs #272