[FEATURE] Replace deprecated license identifier
Current license »GPL-3.0« (https://spdx.org/licenses/GPL-3.0.html)
is deprecated and since it needs to be more specific.

It is replaced with »GPL-3.0-or-later«
(https://spdx.org/licenses/GPL-3.0-or-later.html).

This is backwards compatible (with all existings code and
contributions) whilst being open to future changes of the
GPL license. Which is the most resilient choice for
this project.

Closes #280